store/scope: Switch to uint32 instead of int64
As int64 allows negative values and DB generated code contains
uint32 for the store IDs.

+++ store/scope/type.go
@@ -112,7 +112,7 @@ func (s Type) IsWebSiteOrStore() bool {
 
 // WithID calls MakeTypeID for your convenience. It packs the id into a new value
 // containing the Type and its ID ;-).
-func (s Type) WithID(id int64) TypeID {
+func (s Type) WithID(id uint32) TypeID {
 	return MakeTypeID(s, id)
 }
